Started work on fixing folder structure a bit in diamonds-viewer. Small issue with characterName not overflowing its container.'
