+++ diamonds-viewer/src/blocks/Board/Row.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export default styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  flex-basis: fill;
+`;
